further pr fixes
